fix: stop Gmail watches before source deactivation
